refactor: export types
