+++ src/components/Chart/index.ts
@@ -1,1 +1,2 @@
 export * from './Chart';
+export * from './types';
