+++ src/components/Chart/types.ts
@@ -1,0 +1,19 @@
+import type { Theme } from '@/hooks';
+
+export type ChartConfig = {
+  [key: string]: {
+    icon?: React.ComponentType;
+    label?: React.ReactNode;
+  } & (
+    | {
+        color?: never;
+        theme: {
+          [K in Theme]: string;
+        };
+      }
+    | {
+        color?: string;
+        theme?: never;
+      }
+  );
+};
